DCWL-2003-v2 various clean up and started removing UUID rubbish
