+++ app/uk/gov/hmrc/customs/declarations/information/connectors/ApiSubscriptionFieldsConnector.scala
@@ -22,6 +22,7 @@ import uk.gov.hmrc.customs.declarations.information.config.ConfigService
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
 import uk.gov.hmrc.customs.declarations.information.model.ApiSubscriptionFieldsResponse.format
 import uk.gov.hmrc.customs.declarations.information.model.{ApiSubscriptionFieldsResponse, ApiSubscriptionKey, HasConversationId}
+import uk.gov.hmrc.customs.declarations.information.util.ApiSubscriptionFieldsPath
 import uk.gov.hmrc.http.HttpReads.Implicits.readRaw
 import uk.gov.hmrc.http.{HeaderCarrier, HttpClient}
