+++ app/uk/gov/hmrc/customs/declarations/information/services/declaration/DeclarationVersionService.scala
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.customs.declarations.information.services.declarations
+package uk.gov.hmrc.customs.declarations.information.services.declaration
 
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse.ErrorInternalServerError
 import uk.gov.hmrc.customs.declarations.information.config.ConfigService
 import uk.gov.hmrc.customs.declarations.information.connectors.{ApiSubscriptionFieldsConnector, DeclarationVersionConnector}
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
-import uk.gov.hmrc.customs.declarations.information.services.filters.VersionResponseFilterService
+import uk.gov.hmrc.customs.declarations.information.services.filter.VersionResponseFilterService
 import uk.gov.hmrc.customs.declarations.information.services.UniqueIdsService
 import uk.gov.hmrc.http.HttpResponse
