+++ test/integration/DeclarationSearchConnectorSpec.scala
@@ -49,7 +49,7 @@ class DeclarationSearchConnectorSpec extends IntegrationTestSpec
   }
 
   override protected def beforeEach(): Unit = {
-    when(mockUuidService.uuid()).thenReturn(correlationIdUuid)
+    //when(mockUuidService.uuid()).thenReturn(correlationIdUuid)
   }
 
   override protected def afterEach(): Unit = {
