+++ test/unit/controllers/DeclarationFullControllerSpec.scala
@@ -36,8 +36,8 @@ import uk.gov.hmrc.customs.declarations.information.controllers.DeclarationFullC
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
 import uk.gov.hmrc.customs.declarations.information.model._
 import uk.gov.hmrc.customs.declarations.information.services._
-import uk.gov.hmrc.customs.declarations.information.services.declarations.DeclarationFullService
-import uk.gov.hmrc.customs.declarations.information.services.filters.FullResponseFilterService
+import uk.gov.hmrc.customs.declarations.information.services.declaration.DeclarationFullService
+import uk.gov.hmrc.customs.declarations.information.services.filter.FullResponseFilterService
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import util.ApiSubscriptionFieldsTestData.apiSubscriptionFieldsResponse
 import util.FakeRequests._
