+++ test/unit/controllers/StatusControllerSpec.scala
@@ -36,8 +36,8 @@ import uk.gov.hmrc.customs.declarations.information.controllers.StatusController
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
 import uk.gov.hmrc.customs.declarations.information.model._
 import uk.gov.hmrc.customs.declarations.information.services._
-import uk.gov.hmrc.customs.declarations.information.services.declarations.DeclarationStatusService
-import uk.gov.hmrc.customs.declarations.information.services.filters.StatusResponseFilterService
+import uk.gov.hmrc.customs.declarations.information.services.declaration.DeclarationStatusService
+import uk.gov.hmrc.customs.declarations.information.services.filter.StatusResponseFilterService
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import util.ApiSubscriptionFieldsTestData.apiSubscriptionFieldsResponse
 import util.FakeRequests._
