+++ test/unit/controllers/VersionControllerSpec.scala
@@ -36,8 +36,8 @@ import uk.gov.hmrc.customs.declarations.information.controllers.VersionControlle
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
 import uk.gov.hmrc.customs.declarations.information.model._
 import uk.gov.hmrc.customs.declarations.information.services._
-import uk.gov.hmrc.customs.declarations.information.services.declarations.DeclarationVersionService
-import uk.gov.hmrc.customs.declarations.information.services.filters.VersionResponseFilterService
+import uk.gov.hmrc.customs.declarations.information.services.declaration.DeclarationVersionService
+import uk.gov.hmrc.customs.declarations.information.services.filter.VersionResponseFilterService
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import util.ApiSubscriptionFieldsTestData.apiSubscriptionFieldsResponse
 import util.FakeRequests._
