--- test/unit/services/UuidServiceSpec.scala
@@ -1,36 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package unit.services
-
-import uk.gov.hmrc.customs.declarations.information.services.UuidService
-import util.UnitSpec
-
-class UuidServiceSpec extends UnitSpec {
-
-  private val uuidService = new UuidService
-
-  "UuidService" should {
-    "provide different value on each call" in {
-      val uuid1 = uuidService.uuid()
-      val uuid2 = uuidService.uuid()
-
-      uuid1 should not be uuid2
-      uuid1.toString should not be uuid2.toString
-    }
-  }
-
-}
