+++ test/unit/services/declaration/DeclarationVersionServiceSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package unit.services.declarations
+package unit.services.declaration
 
 import org.mockito.ArgumentMatchers.{eq => meq, _}
 import org.mockito.Mockito.{mock, reset, verify, when}
@@ -32,8 +32,8 @@ import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
 import uk.gov.hmrc.customs.declarations.information.model._
 import ActionBuilderModelHelper._
 import uk.gov.hmrc.customs.declarations.information.services._
-import uk.gov.hmrc.customs.declarations.information.services.declarations.DeclarationVersionService
-import uk.gov.hmrc.customs.declarations.information.services.filters.VersionResponseFilterService
+import uk.gov.hmrc.customs.declarations.information.services.declaration.DeclarationVersionService
+import uk.gov.hmrc.customs.declarations.information.services.filter.VersionResponseFilterService
 import uk.gov.hmrc.customs.declarations.information.xml.BackendVersionPayloadCreator
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import util.ApiSubscriptionFieldsTestData.{apiSubscriptionFieldsResponse, apiSubscriptionFieldsResponseWithEmptyEori, apiSubscriptionFieldsResponseWithNoEori}
